Add session counter to every request
